COMMERCE-1453 Move key to CommerceConstants and internalize admin module for shipping
