+++ commerce-shipping-engine-fixed-web/src/main/resources/META-INF/resources/edit_shipping_option_setting.jsp
@@ -37,7 +37,7 @@ if (commerceShippingFixedOptionRel != null) {
 
 PortletURL shippingMethodsURL = renderResponse.createRenderURL();
 
-shippingMethodsURL.setParameter("commerceAdminModuleKey", ShippingMethodsCommerceAdminModule.KEY);
+shippingMethodsURL.setParameter("commerceAdminModuleKey", commerceAdminModuleKey);
 
 String localizedKey = (commerceShippingFixedOptionRel == null) ? "add-shipping-option-setting" : "edit-shipping-option-setting";
