+++ commerce-shipping-web/bnd.bnd
@@ -1,7 +1,5 @@
 Bundle-Name: Liferay Commerce Shipping Web
 Bundle-SymbolicName: com.liferay.commerce.shipping.web
 Bundle-Version: 2.0.5
-Export-Package:\
-	com.liferay.commerce.shipping.web.admin,\
-	com.liferay.commerce.shipping.web.servlet.taglib.ui
+Export-Package: com.liferay.commerce.shipping.web.servlet.taglib.ui
 Web-ContextPath: /commerce-shipping-web
