+++ commerce-shipping-web/build.gradle
@@ -3,7 +3,6 @@ apply plugin: "com.liferay.lang.merger"
 dependencies {
 	compileOnly group: "com.liferay", name: "com.liferay.frontend.taglib", version: "3.0.0"
 	compileOnly group: "com.liferay.commerce", name: "com.liferay.commerce.admin.api", version: "2.0.0"
-	compileOnly group: "com.liferay.commerce", name: "com.liferay.commerce.api", version: "9.3.0"
 	compileOnly group: "com.liferay.commerce", name: "com.liferay.commerce.inventory.api", version: "1.0.0"
 	compileOnly group: "com.liferay.commerce", name: "com.liferay.commerce.item.selector.api", version: "2.0.0"
 	compileOnly group: "com.liferay.commerce", name: "com.liferay.commerce.service", version: "5.2.0"
@@ -12,4 +11,5 @@ dependencies {
 	compileOnly group: "javax.portlet", name: "portlet-api", version: "3.0.0"
 	compileOnly group: "javax.servlet", name: "javax.servlet-api", version: "3.0.1"
 	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
+	compileOnly project(":apps:commerce:commerce-api")
 }
