+++ commerce-shipping-web/src/main/resources/META-INF/resources/init.jsp
@@ -29,7 +29,7 @@ page import="com.liferay.commerce.exception.CommerceShippingMethodNameException"
 page import="com.liferay.commerce.model.CommerceAddressRestriction" %><%@
 page import="com.liferay.commerce.model.CommerceCountry" %><%@
 page import="com.liferay.commerce.model.CommerceShippingMethod" %><%@
-page import="com.liferay.commerce.shipping.web.admin.ShippingMethodsCommerceAdminModule" %><%@
+page import="com.liferay.commerce.shipping.web.internal.admin.ShippingMethodsCommerceAdminModule" %><%@
 page import="com.liferay.commerce.shipping.web.internal.display.context.CommerceShippingMethodRestrictionsDisplayContext" %><%@
 page import="com.liferay.commerce.shipping.web.internal.display.context.CommerceShippingMethodsDisplayContext" %><%@
 page import="com.liferay.commerce.shipping.web.servlet.taglib.ui.CommerceShippingScreenNavigationConstants" %><%@
