--- commerce-shipping-web/src/main/resources/com/liferay/commerce/shipping/web/admin/packageinfo
@@ -1,1 +1,0 @@
-version 1.1.0
